update sublimelinter-flow to use flow's json output
Commit is mostly documentation, but i think it is easier to follow
what is going on and it's pretty straightforward in terms of how to
improve logging or what have you.

I was having trouble using sublimelinter-flow earlier, i think maybe
the old format was colliding against the regex, but this seems to
address that. Tested against a couple of in-progress files.