Only lint if @flow pragma is present
doesn't intelligently check for it in the first comment like flow will,
but this check seems to be good enough to halt spurious flow checks.

additionally, uses the view's contents piped to stdin to check for
errors rather than using a tempfile

Uses newer json output to provide better errors in flow output.