Add support for flow coverage.
If a user enables coverage explicitly in the linter settings or inline,
source code not covered by flow will be highlighted as a warning.

Since flow implements type checks and coverage in different commands,
the linter is invoked twice. Flow runs as a daemon/server, i.e. it only
adds negligible overhead.

Check and coverage results are combined into a single JSON string within
the `run` method, and parsed as JSON array in `find_errors`.